requirements.txt
fastapi 
uvicorn 
langchain 
llama-index 
requests 
beautifulsoup4 
supabase 
python-dotenv 
pydantic 
aiohttp 
openai
transformers
streamlit
faiss-cpu
selenium
lxml
html5lib
streamlit
openai
tiktoken